requirements.txt
absl-py==2.4.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astunparse==1.6.3
bcrypt==3.2.2
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==46.0.7
deepface==0.0.99
ecdsa==0.19.2
fastapi==0.136.1
filelock==3.29.0
fire==0.7.1
Flask==3.1.3
flask-cors==6.0.2
flatbuffers==25.12.19
gast==0.7.0
gdown==6.0.0
google-pasta==0.2.0
greenlet==3.4.0
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
h5py==3.14.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.14.0
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.6
lightphe==0.0.24
lz4==4.4.5
Mako==1.3.11
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
numpy==2.4.4
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opt_einsum==3.4.0
optree==0.19.0
packaging==26.1
pandas==3.0.2
passlib==1.7.4
pgvector==0.4.2
pillow==12.2.0
protobuf==7.34.1
psycopg2-binary==2.9.12
pyasn1==0.6.3
pycparser==3.0
pydantic==2.13.3
pydantic-settings==2.14.0
pydantic_core==2.46.3
Pygments==2.20.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
requests==2.33.1
retina-face==0.0.17
rich==15.0.0
rsa==4.9.1
setuptools==82.0.1
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
starlette==1.0.0
sympy==1.14.0
tensorflow==2.21.0
termcolor==3.3.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.46.0
Werkzeug==3.1.8
wheel==0.47.0
wrapt==2.1.2
